Fix FindMigrations panic with missing migrations
If a migration was missing and the directory contained only valid
migration files, FindMigrations would panic with an index out of range
error instead of returning an error indicating a missing migration.

This was hidden in the tests because the test data included a
non-migration file that added an extra element to the array.