Disable prepared statements when run from CLI
Apparently, when run through a connection pooler, the prepared
statements from one run may persist to the next run because the
underlying connection is not closed and it retains the prepared
statements. This would cause the next run to fail because the prepared
statement names are deterministic and would conflict with the already
existing prepared statements.

https://github.com/jackc/tern/issues/100