Fix fallback config confusion
If the sslmode was blank then some of the fallback config was not being
set properly. This was because blank is treated as "prefer" by pgx
(which follows libpq's behavior). But tern parsing parsing the config
in several places and then trying to fix the fallback config. It missed
the default "prefer" case.

The solution was to avoid the multiple parsing and fallback config logic
entirely. Instead, we set the PG* environment variables from the CLI
options and config file before parsing the conn string.

https://github.com/jackc/tern/issues/38